UPDATE: add editor classes to open and edit file for note
Note that supported editors should be able to wait for user editing
before returning it control back to our application. Otherwise, that
editor cannot be supported.

For example, VSCode will return an exit code right after opening a
file by the following command:
```bash
$ code FILE_NAME
```

To make it able to wait for editing, we should add an argument
`-w/--wait`:
```bash
$ code --wait FILE_NAME
```