MAINT: rename returned value from `Notebook.get_or_create()`
According to the documentation of `peewee`, the second returned
value of `Model.get_or_create()` is a boolean indicating if a new
object was created. However, it was named `exists` in previous
revisions and it does not make a correct expression. So that it is
renamed as `not_existed` now.