MAINT: change `create_time` and `update_time` of `qnote.object.Notebook` to object property
In order to make it easier to convert data obtained from database
to `qnote.object.Notebook`, we change those fields above-mentioned
to object property with setters. So that values with different
types can be handled automatically while setting those properties.

The reason why we made this change is that those values retrieved
from SQLite database in the type of `TimestampField` will be
converted to `datetime.datetime` object automatically. However,
what we initially expect is `int` for those timestamp.