BUG: failed to get notes with no tag
If a note does not tagged, it won't be shown after executing
`qnote list` command.

The reason why it is not able to be retrieved is that the argument
`join_type` of `peewee.Model.join()` is `JOIN.INNER` by default.

Take a look at the query in `sqlite.storer.get_notes_from_notebook()`:

  ```python
  Note
  .select()
  .join(NoteToNotebook)
  .join(Notebook)
  # ...
  .join(NoteToTag, on=(Note.id == NoteToTag.note_id))
  .join(Tag)
  # ...
  ```

Since we can only get those items existing in both tables while
using a inner join, those notes without tags in this query will
be excluded by the last 2 `join` operation. In brief, it's a
intersection instead of union between "Note" and "Tag" tables.

To solve this issue, we have to perform a left outer join to leave
those notes without tags existing in query. That's why we specified
`join_type` to LEFT_OUTER in the last two `join` operation.