MAINT: rewrite code related to `utils.text.show_note()`
In previous revisions, content to be printed and redirected to a
pager is required to be dumped into a tempfile. But it's not
necessary to do so, we can use "PIPE" to redict content to a pager
instead.

Take `less` as an example. Usually, we can use it to view the
content of a file by the following command:
```bash
$ less FILE_NAME
```

Or we can use pipe operator to redirect the content to `less`:
```bash
$ cat FILE_NAME | less
```

Difference between those 2 example above is that `less` doesn't
need to read a file directly. With this advantage, we can avoid
writing content to a tempfile before calling `less` to read it.

See also:
https://superuser.com/questions/81679