BUG: failed to get notes without tags by uuid
The cause is the same as the one mentioned in commit 01a8088.

Once a note is not realted to any tag, it will be excluded when
the selection query is followed by a "JOIN" operation with `Tag`
or `NoteToTag` tables.

To solve this issue, we have to specify `join_type` to `LEFT_OUTER`.