MAINT: set argument `notebook` as positional argument in command `qnote move` and some minor revisions
Though the original propose of setting `notebook` as an optional
argument is to make it can be specified clearly while entering a
command. However, it makes itself not able to be parsed correctly
when we are going to execute subcommand `qnote move selected`.

Since optional arguments defined in (parent) parser are always
parsed first, there is currently no easy approach to change this
behavior without rewriting `ArgumentParser.parse_known_args()`
method. (see also the part of code in it related to
`consume_optional()`, where `consume_optional()` is called before
`consume_positionals()`)

Other revisions:
- fix typo in `NotebookCommand`
- In "commands/remove.py", `ARG_SUPPRESS` is imported from
  `qnote.cli.parser` instead of `argparse`.
- In `NoteManager`, statement of `get_storer()` in some methods
  are repositioned in order to make it fail fast when there is any
  checks failed without establishing a connection to database.